+++ test-app/db/migrations/1751044835745-create-polymorphic-workout.ts
@@ -4,7 +4,7 @@ export async function up(db: Kysely<any>): Promise<void> {
   await db.schema
     .createTable('polymorphic_workouts')
     .addColumn('id', 'bigserial', col => col.primaryKey())
-    .addColumn('name', 'varchar(255)', col => col.notNull())
+    .addColumn('name', 'varchar(255)')
     .addColumn('created_at', 'timestamp', col => col.notNull())
     .addColumn('updated_at', 'timestamp', col => col.notNull())
     .execute()